initial commit and stuff
welcome to the world MANSI! MANSI is multiresolution atomic norm system
identification, which aims to perform sysid given input data (noisy
impulse/step response etc.). Normally, a fine epsilon net is needed to
ensure convergence. MANSI aims to change this, by only zooming into
regions which are previously found to be important. This should improve
the sparsity properties, and also be much faster.

This commit focuses on ANSI, which is not multiresolution. It recasts
sysid into optimization over the L1 ball, which has been solved in the
Elastic Net code. More work will be done on this.