further development, and fixed some bugs
added tons of plots, away steps, and some other features. Fixed a
critical bug with pole_matrix which screwed up the scaling. There still
might be some bugs with scaling, and the best call might be to just
normalize the columns instead of calculating a scaling value. The code
itself does not work as there is something weird in the gradient
calculation, a ground truth pole of 0.5 will have the algorithm start
off at the pole by 0.3. More research needs to be done to see what the
hell is going on.