found transfer function (fitted and reduced)
found the values of the residues r from the sin/cos coefficients c.
There still might be a bit of trouble here, had to use some hacky matlab
in order to get it working. This allows for combining the partial
fraction terms into a single transfer function.

Using residue() to get the transfer function caused great numeric
difficulties, where the known pole locations were completely corrupted.
I extracted the zero-finding routine from residue and placed it into
zeros_from_poles. For sparse-fitted systems, the pole-zero cancellations
are quite observable. As the order of the fitted model increases, the
zeros tend to spread out and go beyond the unit circle, no longer
cancelling their respective poles. This is a numerical issue, and I am
not sure how to possibly deal with this.